refactor(connect): rename blockchainValidateEvmRpcUrl to blockchainEvmRpcGetChainId
