+++ packages/connect-common/src/types/api/blockchain/blockchainEvmRpcGetChainId.ts
