+++ packages/connect-common/src/types/api/blockchain/index.ts
@@ -4,6 +4,7 @@ import { Type } from '@trezor/schema-utils';
 import type { blockchainDisconnect } from './blockchainDisconnect';
 import type { blockchainEstimateFee } from './blockchainEstimateFee';
 import type { blockchainEvmRpcCall } from './blockchainEvmRpcCall';
+import type { blockchainValidateEvmRpcUrl } from './blockchainEvmRpcGetChainId';
 import type { blockchainGetAccountBalanceHistory } from './blockchainGetAccountBalanceHistory';
 import type { blockchainGetContractInfo } from './blockchainGetContractInfo';
 import type { blockchainGetCurrentFiatRates } from './blockchainGetCurrentFiatRates';
@@ -15,7 +16,6 @@ import type { blockchainSubscribe } from './blockchainSubscribe';
 import type { blockchainSubscribeFiatRates } from './blockchainSubscribeFiatRates';
 import type { blockchainUnsubscribe } from './blockchainUnsubscribe';
 import type { blockchainUnsubscribeFiatRates } from './blockchainUnsubscribeFiatRates';
-import type { blockchainValidateEvmRpcUrl } from './blockchainValidateEvmRpcUrl';
 import type { pushTransaction } from './pushTransaction';
 
 // Blockchain backend operations (no device needed)
