+++ packages/connect/src/api/index.ts
@@ -8,6 +8,7 @@ export { default as backupDevice } from './backupDevice';
 export { default as bleUnpair } from './bleUnpair';
 export { default as blockchainDisconnect } from './blockchainDisconnect';
 export { default as blockchainEstimateFee } from './blockchainEstimateFee';
+export { default as blockchainValidateEvmRpcUrl } from './blockchainEvmRpcGetChainId';
 export { default as blockchainGetAccountBalanceHistory } from './blockchainGetAccountBalanceHistory';
 export { default as blockchainGetCurrentFiatRates } from './blockchainGetCurrentFiatRates';
 export { default as blockchainGetContractInfo } from './blockchainGetContractInfo';
@@ -20,7 +21,6 @@ export { default as blockchainSubscribe } from './blockchainSubscribe';
 export { default as blockchainSubscribeFiatRates } from './blockchainSubscribeFiatRates';
 export { default as blockchainUnsubscribe } from './blockchainUnsubscribe';
 export { default as blockchainUnsubscribeFiatRates } from './blockchainUnsubscribeFiatRates';
-export { default as blockchainValidateEvmRpcUrl } from './blockchainValidateEvmRpcUrl';
 export { default as changeLanguage } from './changeLanguage';
 export { default as changePin } from './changePin';
 export { default as changeWipeCode } from './changeWipeCode';
